Add interface for spanner
