+++ slasher/detection/attestations/spanner_test.go
@@ -7,7 +7,11 @@ import (
 	"testing"
 
 	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
+	"github.com/prysmaticlabs/prysm/slasher/detection/attestations/iface"
+	"github.com/prysmaticlabs/prysm/slasher/detection/attestations/types"
 )
+
+var _ = iface.SpanDetector{&SpanDetector{})
 
 func TestSpanDetector_DetectSlashingForValidator(t *testing.T) {
 	type testStruct struct {
@@ -229,8 +233,8 @@ func TestSpanDetector_DetectSlashingForValidator(t *testing.T) {
 				t.Fatalf("Did not want validator to be slashed but found slashable offense: %v", res)
 			}
 			if tt.shouldSlash {
-				want := &DetectionResult{
-					Kind:           SurroundVote,
+				want := &types.DetectionResult{
+					Kind:           types.SurroundVote,
 					SlashableEpoch: tt.slashableEpoch,
 				}
 				if !reflect.DeepEqual(res, want) {
@@ -335,8 +339,8 @@ func TestSpanDetector_DetectSlashingForValidator_MultipleValidators(t *testing.T
 					t.Fatalf("Did not want validator to be slashed but found slashable offense: %v", res)
 				}
 				if tt.shouldSlash[valIdx] {
-					want := &DetectionResult{
-						Kind:           SurroundVote,
+					want := &types.DetectionResult{
+						Kind:           types.SurroundVote,
 						SlashableEpoch: tt.slashableEpochs[valIdx],
 					}
 					if !reflect.DeepEqual(res, want) {
