+++ slasher/detection/attestations/types/types.go
@@ -1,0 +1,25 @@
+package types
+
+// DetectionKind defines an enum type that
+// gives us information on the type of slashable offense
+// found when analyzing validator min-max spans.
+type DetectionKind int
+
+const (
+	// DoubleVote denotes a slashable offense in which
+	// a validator cast two conflicting attestations within
+	// the same target epoch.
+	DoubleVote DetectionKind = iota
+	// SurroundVote denotes a slashable offense in which
+	// a validator surrounded or was surrounded by a previous
+	// attestation created by the same validator.
+	SurroundVote
+)
+
+// DetectionResult tells us the kind of slashable
+// offense found from detecting on min-max spans +
+// the slashable epoch for the offense.
+type DetectionResult struct {
+	Kind           DetectionKind
+	SlashableEpoch uint64
+}
